Update Dependencies (#47)
* initial commit

* ui updates & fix subsetting for ugrid

* update python in circleci

---------
