+++ app.py
@@ -42,9 +42,8 @@ app.add_middleware(
 app.mount("/", SPAStaticFiles(directory="./viewer/dist", html=True), name="viewer")
 app.root_path = settings.root_path
 
-
 if __name__ == "__main__":
     import uvicorn
 
     # When run directly, run in debug mode
-    uvicorn.run("app:app", host="0.0.0.0", port=8090, reload=True, log_level="debug")
+    uvicorn.run("app:app", host="0.0.0.0", port=8090, reload=False, log_level="debug")
