+++ requirements.txt
@@ -1,42 +1,44 @@
-cachey==0.2.1
-cf-xarray==0.9.0
-cfgrib==0.9.10.2
-cftime==1.6.2
-dask==2024.4.2
+cachey~=0.2.1 # last updated 03/20
+cf-xarray~=0.10.0
+cfgrib~=0.9.15.0
+cftime~=1.6.4.post1
+# dask/distributed are pinned because newer versions break netcdf requests
+dask==2024.11.2
 dask_gateway==2024.1.0
-distributed==2024.4.2
-fastapi==0.110.2
-gunicorn==22.0.0
-h11==0.14.0
-h5netcdf==1.3.0
-h5py==3.11.0
-matplotlib==3.6.1
-mercantile==1.2.1
-netCDF4==1.6.5
-numcodecs==0.12.1
-numpy==1.26.4
-pandas==2.2.2
-Pillow==10.3.0
-pluggy==1.5.0
-pydantic==2.7.1
-pydantic-settings==2.2.1
-pyproj==3.6.1
-python-dateutil==2.9.0.post0
-pytz==2024.1
-rasterio==1.3.10
-rioxarray==0.15.5
-fsspec==2024.3.1
-s3fs==2024.3.1
-uvicorn==0.29.0
-xarray==2024.3.0
-xpublish@git+https://github.com/xpublish-community/xpublish.git@a4c7d179410908d7a9268cf5e285ecc552b8675d
-zarr==2.17.2
-scipy==1.13.0
-kerchunk@git+https://github.com/fsspec/kerchunk@8d8e84c619a63d4372b33b05c9bfe35fdcb6e61d
-xpublish-opendap@git+https://github.com/xpublish-community/xpublish-opendap@6020c0464f9b7f872f4307dab9d7c97d5d981387
-xpublish-wms@git+https://github.com/xpublish-community/xpublish-wms@9574a71405e43c479b7ebcfa1e5d0def04c598d5
-xpublish-edr@git+https://github.com/xpublish-community/xpublish-edr@019e53acd2e0ad5a1d909d1acfe9863f2e90e51b
-opendap-protocol<1.2.0
-xarray-subset-grid@git+https://github.com/asascience-open/xarray-subset-grid@2de3f6f3ecd8de26f3c6b0b73ce77f9cd362f4f2
-redis-fsspec-cache@git+https://github.com/mpiannucci/redis-fsspec-cache.git@c5f241f113964ec844dbdd48f9eae6119290d1fa
-redis==5.0.4
+distributed==2024.11.2
+fastapi~=0.115.7
+fsspec~=2024.12.0
+gunicorn~=23.0.0
+h11~=0.14.0 # last updated 09/22
+h5netcdf~=1.5.0
+h5py~=3.12.1
+kerchunk~=0.2.7
+matplotlib~=3.10.0
+mercantile~=1.2.1 # last updated 04/21
+netCDF4~=1.7.2
+numcodecs~=0.15.0
+numpy~=1.26.4
+opendap-protocol~=1.1.1 # last updated 09/21
+pandas~=2.2.3
+pillow~=11.1.0
+pluggy~=1.5.0
+pydantic~=2.10.6
+pydantic-settings~=2.7.1
+pyproj~=3.7.0
+python-dateutil~=2.9.0.post0
+pytz~=2024.2
+rasterio~=1.4.3
+redis~=5.2.1
+rioxarray~=0.18.2
+s3fs~=2024.12.0
+scipy~=1.15.1
+setuptools~=75.8.0
+uvicorn~=0.34.0
+xarray~=2025.1.2
+zarr~=2.18.4
+redis-fsspec-cache@git+https://github.com/mpiannucci/redis-fsspec-cache@main
+xarray-subset-grid@git+https://github.com/asascience-open/xarray-subset-grid@main
+xpublish@git+https://github.com/xpublish-community/xpublish@main
+xpublish-opendap@git+https://github.com/xpublish-community/xpublish-opendap@main
+xpublish-wms@git+https://github.com/xpublish-community/xpublish-wms@main
+xpublish-edr@git+https://github.com/xpublish-community/xpublish-edr@main
