+++ viewer/package.json
@@ -10,30 +10,30 @@
         "preview": "vite preview"
     },
     "dependencies": {
-        "@mapbox/mapbox-gl-draw": "^1.4.3",
-        "@tanstack/query-core": "^5.32.1",
-        "@tanstack/react-query": "^5.32.1",
-        "fast-xml-parser": "^4.4.1",
+        "@mapbox/mapbox-gl-draw": "^1.5.0",
+        "@tanstack/query-core": "^5.65.0",
+        "@tanstack/react-query": "^5.65.1",
+        "fast-xml-parser": "^4.5.1",
         "he": "^1.2.0",
-        "maplibre-gl": "^2.4.0",
-        "material-icons": "^1.12.2",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
-        "react-router-dom": "^6.23.0",
-        "wonka": "^6.3.4"
+        "maplibre-gl": "^5.1.0",
+        "material-icons": "^1.13.13",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
+        "react-router-dom": "^7.1.3"
     },
     "devDependencies": {
-        "@types/he": "^1.1.2",
-        "@types/mapbox__mapbox-gl-draw": "^1.4.6",
-        "@types/react": "^18.0.24",
-        "@types/react-dom": "^18.0.8",
+        "@tailwindcss/postcss": "^4.0.0",
+        "@types/he": "^1.2.3",
+        "@types/mapbox__mapbox-gl-draw": "^1.4.8",
+        "@types/react": "^18.3.18",
+        "@types/react-dom": "^18.3.5",
         "@types/react-router-dom": "^5.3.3",
-        "@vitejs/plugin-react": "^2.2.0",
-        "autoprefixer": "^10.4.13",
-        "postcss": "^8.4.19",
-        "tailwindcss": "^3.2.4",
-        "typescript": "^4.6.4",
-        "vite": "^3.2.7"
+        "@vitejs/plugin-react": "^4.3.4",
+        "autoprefixer": "^10.4.20",
+        "postcss": "^8.5.1",
+        "tailwindcss": "^4.0.0",
+        "typescript": "^5.7.3",
+        "vite": "^6.0.11"
     },
     "prettier": {
         "singleQuote": true,
